refactor(conversations): satisfy architecture ratchet
