+++ lemma-backend/app/modules/agent/api/controllers/conversation_controller.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import anyio
-from collections.abc import AsyncGenerator, Iterable
+from collections.abc import AsyncGenerator, Awaitable, Callable, Iterable
 from uuid import UUID
 
 from fastapi import APIRouter, Depends, HTTPException, Query, Request, status
@@ -43,6 +43,7 @@ from app.modules.agent.domain.errors import (
     ConversationNotFoundError,
 )
 from app.modules.agent.domain.value_objects import (
+    AgentRunStartResult,
     ConversationAgentSelection,
     ConversationStatus,
     ConversationType,
@@ -52,9 +53,11 @@ from app.modules.agent.infrastructure.repositories import (
     AgentRepository,
     ConversationRepository,
 )
-from app.modules.agent.services.conversation_service import ConversationService
+from app.modules.agent.services.conversation_retry_service import (
+    ConversationRetryService,
+)
 from app.composition.authorization import create_authorization_service
-from app.composition.agent_usage import build_usage_service, UsageLimitExceededError
+from app.composition.agent_usage import build_usage_service
 
 logger = get_logger(__name__)
 
@@ -64,14 +67,64 @@ router = APIRouter(
 )
 
 
-def _build_conversation_service(uow) -> ConversationService:
-    return ConversationService(
+def _build_conversation_service(uow) -> ConversationRetryService:
+    return ConversationRetryService(
         uow=uow,
         conversation_repository=ConversationRepository(uow),
         agent_repository=AgentRepository(uow),
         authorization_service=create_authorization_service(uow),
         usage_service=build_usage_service(uow),
     )
+
+
+async def _start_and_stream_run(
+    *,
+    channel_service: ChannelServiceDep,
+    conversation_id: UUID,
+    start_run: Callable[[], Awaitable[AgentRunStartResult]],
+) -> StreamingResponse:
+    async def close_subscription(
+        exc_type=None,
+        exc=None,
+        traceback=None,
+    ) -> None:
+        try:
+            with anyio.CancelScope(shield=True):
+                await subscription.__aexit__(exc_type, exc, traceback)
+        except Exception:
+            return
+
+    subscription = channel_service.subscribe([conversation_channel(conversation_id)])
+    iterator = await subscription.__aenter__()
+    try:
+        result = await start_run()
+    except (AgentNotFoundError, ConversationNotFoundError) as exc:
+        await close_subscription(type(exc), exc, exc.__traceback__)
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND) from exc
+    except BaseException as exc:
+        await close_subscription(type(exc), exc, exc.__traceback__)
+        raise
+
+    async def event_generator() -> AsyncGenerator[str, None]:
+        try:
+            async for chunk in iter_subscription(iterator, result.agent_run_id):
+                yield chunk
+        except Exception:
+            logger.error(
+                "agent.conversation_controller.agent_realtime_subscription.failed",
+                conversation_id=str(conversation_id),
+                agent_run_id=str(result.agent_run_id),
+                exc_info=True,
+            )
+            yield encode_stream_chunk(
+                event_type="error",
+                data="Realtime stream interrupted. Reconnect to continue.",
+                agent_run_id=result.agent_run_id,
+            )
+        finally:
+            await close_subscription()
+
+    return StreamingResponse(event_generator(), media_type="text/event-stream")
 
 
 def _parse_metadata_filters(
@@ -392,61 +445,24 @@ async def send_message(
     # dependencies (and their pooled connection) alive for the whole SSE stream,
     # which pins one DB connection per in-flight stream. Here the connection is
     # released the moment add_user_message_and_start_run commits, before streaming.
-    async def close_subscription(
-        exc_type=None,
-        exc=None,
-        traceback=None,
-    ) -> None:
-        try:
-            with anyio.CancelScope(shield=True):
-                await subscription.__aexit__(exc_type, exc, traceback)
-        except Exception:
-            return
-
-    subscription = channel_service.subscribe([conversation_channel(conversation_id)])
-    iterator = await subscription.__aenter__()
-    try:
+    async def start_run() -> AgentRunStartResult:
         async with pod_context_scope(
             uow_factory, request=request, user_id=user.id, pod_id=pod_id
         ) as scope:
             service = _build_conversation_service(scope.uow)
-            result = await service.add_user_message_and_start_run(
+            return await service.add_user_message_and_start_run(
                 conversation_id=conversation_id,
                 user_id=user.id,
                 content=data.content,
                 pod_id=pod_id,
                 message_metadata=data.metadata,
             )
-    except (AgentNotFoundError, ConversationNotFoundError) as exc:
-        await close_subscription(type(exc), exc, exc.__traceback__)
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND) from exc
-    except UsageLimitExceededError as exc:
-        await close_subscription(type(exc), exc, exc.__traceback__)
-        raise
-    except BaseException as exc:
-        await close_subscription(type(exc), exc, exc.__traceback__)
-        raise
 
-    async def event_generator() -> AsyncGenerator[str, None]:
-        try:
-            async for chunk in iter_subscription(iterator, result.agent_run_id):
-                yield chunk
-        except Exception:
-            logger.error(
-                "agent.conversation_controller.agent_realtime_subscription.failed",
-                conversation_id=str(conversation_id),
-                agent_run_id=str(result.agent_run_id),
-                exc_info=True,
-            )
-            yield encode_stream_chunk(
-                event_type="error",
-                data="Realtime stream interrupted. Reconnect to continue.",
-                agent_run_id=result.agent_run_id,
-            )
-        finally:
-            await close_subscription()
-
-    return StreamingResponse(event_generator(), media_type="text/event-stream")
+    return await _start_and_stream_run(
+        channel_service=channel_service,
+        conversation_id=conversation_id,
+        start_run=start_run,
+    )
 
 
 @router.post(
@@ -468,56 +484,22 @@ async def retry_failed_run(
     request: Request,
     uow_factory: UnitOfWorkFactory = Depends(get_uow_factory),
 ) -> StreamingResponse:
-    async def close_subscription(
-        exc_type=None,
-        exc=None,
-        traceback=None,
-    ) -> None:
-        try:
-            with anyio.CancelScope(shield=True):
-                await subscription.__aexit__(exc_type, exc, traceback)
-        except Exception:
-            return
-
-    subscription = channel_service.subscribe([conversation_channel(conversation_id)])
-    iterator = await subscription.__aenter__()
-    try:
+    async def start_run() -> AgentRunStartResult:
         async with pod_context_scope(
             uow_factory, request=request, user_id=user.id, pod_id=pod_id
         ) as scope:
             service = _build_conversation_service(scope.uow)
-            result = await service.retry_failed_run(
+            return await service.retry_failed_run(
                 conversation_id=conversation_id,
                 user_id=user.id,
                 pod_id=pod_id,
             )
-    except (AgentNotFoundError, ConversationNotFoundError) as exc:
-        await close_subscription(type(exc), exc, exc.__traceback__)
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND) from exc
-    except BaseException as exc:
-        await close_subscription(type(exc), exc, exc.__traceback__)
-        raise
 
-    async def event_generator() -> AsyncGenerator[str, None]:
-        try:
-            async for chunk in iter_subscription(iterator, result.agent_run_id):
-                yield chunk
-        except Exception:
-            logger.error(
-                "agent.conversation_controller.agent_realtime_subscription.failed",
-                conversation_id=str(conversation_id),
-                agent_run_id=str(result.agent_run_id),
-                exc_info=True,
-            )
-            yield encode_stream_chunk(
-                event_type="error",
-                data="Realtime stream interrupted. Reconnect to continue.",
-                agent_run_id=result.agent_run_id,
-            )
-        finally:
-            await close_subscription()
-
-    return StreamingResponse(event_generator(), media_type="text/event-stream")
+    return await _start_and_stream_run(
+        channel_service=channel_service,
+        conversation_id=conversation_id,
+        start_run=start_run,
+    )
 
 
 @router.get(
